chore: split code into mixins
