+++ packages/shared/widget-plugin-filtering/src/controllers/picker/RefTagPickerController.ts
@@ -1,153 +1,13 @@
-import { useCombobox, UseComboboxProps, useMultipleSelection, UseMultipleSelectionProps } from "downshift";
-import { action, autorun, computed, makeObservable, observable } from "mobx";
-import { disposeFx } from "../../mobx-utils";
-import { OptionWithState } from "../../typings/OptionWithState";
 import { RefBaseController, RefBaseControllerProps } from "./RefBaseController";
+import { TagPickerControllerMixin } from "./mixins/TagPickerControllerMixin";
 
-interface Props extends RefBaseControllerProps {
-    inputPlaceholder?: string;
-}
-
-export class RefTagPickerController extends RefBaseController {
-    private touched = false;
-    inputValue = "";
-    inputPlaceholder: string;
-
-    constructor(props: Props) {
+export class RefTagPickerController extends TagPickerControllerMixin(RefBaseController) {
+    constructor(props: RefBaseControllerProps) {
         super(props);
         this.inputPlaceholder = props.inputPlaceholder ?? "Search";
-
-        makeObservable<this, "touched" | "setTouched">(this, {
-            inputValue: observable,
-            touched: observable,
-            setTouched: action,
-            selectedIndex: computed,
-            setInputValue: action,
-            handleBlur: action,
-            handleClear: action,
-            handleFocus: action
-        });
-    }
-
-    setup(): () => void {
-        const [disposers, dispose] = disposeFx();
-        disposers.push(autorun(...this.searchSyncFx()));
-        disposers.push(super.setup());
-        return dispose;
-    }
-
-    searchSyncFx(): Parameters<typeof autorun> {
-        const effect = (): void => {
-            const { touched, inputValue } = this;
-            if (touched) {
-                this.filterStore.search.setBuffer(inputValue);
-            } else {
-                this.filterStore.search.clear();
-            }
-        };
-
-        return [effect];
-    }
-
-    get selectedIndex(): number {
-        const index = this.filterStore.options.findIndex(option => option.selected);
-        return Math.max(index, 0);
-    }
-
-    get selectedOptions(): OptionWithState[] {
-        return this.filterStore.selectedOptions;
-    }
-
-    private setTouched(value: boolean): void {
-        this.touched = value;
-    }
-
-    setInputValue(value: string): void {
-        this.inputValue = value;
     }
 
     handleFocus = (): void => {
         this.filterStore.setFetchReady(true);
-    };
-
-    handleBlur = (): void => {
-        this.setTouched(false);
-        this.setInputValue("");
-        this.filterStore.search.clear();
-    };
-
-    handleClear = (): void => {
-        this.setTouched(false);
-        this.setInputValue("");
-        this.filterStore.clear();
-    };
-
-    useComboboxProps = (): UseComboboxProps<OptionWithState> => {
-        const props: UseComboboxProps<OptionWithState> = {
-            items: this.options,
-            itemToKey: item => item?.value,
-            itemToString: item => item?.caption ?? "",
-            inputValue: this.inputValue,
-            defaultHighlightedIndex: this.selectedIndex,
-            onInputValueChange: changes => {
-                // Blur is handled by handleBlur;
-                if (changes.type === useCombobox.stateChangeTypes.InputBlur) {
-                    return;
-                }
-                if (changes.type === useCombobox.stateChangeTypes.InputChange) {
-                    this.setTouched(true);
-                }
-                this.setInputValue(changes.inputValue);
-            },
-            onSelectedItemChange: ({ selectedItem, type }) => {
-                if (
-                    type === useCombobox.stateChangeTypes.InputBlur ||
-                    type === useCombobox.stateChangeTypes.InputKeyDownEscape ||
-                    !selectedItem
-                ) {
-                    return;
-                }
-                this.filterStore.toggle(selectedItem.value);
-            },
-            stateReducer(state, { changes, type }) {
-                switch (type) {
-                    case useCombobox.stateChangeTypes.InputKeyDownEnter:
-                    case useCombobox.stateChangeTypes.ItemClick:
-                        return {
-                            ...changes,
-                            isOpen: true,
-                            highlightedIndex: state.highlightedIndex,
-                            inputValue: state.inputValue
-                        };
-                    default:
-                        return {
-                            ...changes,
-                            highlightedIndex: changes.inputValue !== state.inputValue ? 0 : changes.highlightedIndex
-                        };
-                }
-            }
-        };
-        return props;
-    };
-
-    useMultipleSelectionProps = (): UseMultipleSelectionProps<OptionWithState> => {
-        const props: UseMultipleSelectionProps<OptionWithState> = {
-            selectedItems: this.selectedOptions,
-            onStateChange: ({ selectedItems: newSelectedItems, type }) => {
-                newSelectedItems ??= [];
-                switch (type) {
-                    case useMultipleSelection.stateChangeTypes.SelectedItemKeyDownBackspace:
-                    case useMultipleSelection.stateChangeTypes.SelectedItemKeyDownDelete:
-                    case useMultipleSelection.stateChangeTypes.DropdownKeyDownBackspace:
-                    case useMultipleSelection.stateChangeTypes.FunctionRemoveSelectedItem:
-                        this.filterStore.setSelected(newSelectedItems.map(item => item.value));
-                        break;
-                    default:
-                        break;
-                }
-            }
-        };
-
-        return props;
     };
 }
