+++ packages/shared/widget-plugin-filtering/src/typings/type-utils.ts
@@ -1,3 +1,5 @@
 export type FnName<T> = T extends { name: infer Name } ? Name : never;
 
 export type Dispose = () => void;
+
+export type GConstructor<T = unknown> = new (...args: any[]) => T;
